Rollback trigger: test failure during validation.
